Upload files as sample stage artefacts.
The files are expected to be very large (on the order of Gb), so they
are uploaded in parts in an effort to make the process more robust and
retries more efficient in the event of failures.

Currently the multi-part upload isn't quite as efficient as it could as
the parts are very small (relative to the overall file size).  Larger
part sizes tend to cause network errors during the upload.  I suspect
that ths is due to the server sending data back to the client during the
upload; the Keep-Alive timeout is tripped and causes the client to think
that the connection has dropped.

This is only a hypothesis; fruther investigation is required.