Add support for cancelling uploads.
Besides the obvious addition of the button, the most significant change
is actually to the file completion handling.  This version of Resumable
triggers file completion events for each simultaneous upload when an
upload for a file in cancelled.  As a result, the completion event is
now triggered (more correctly, especially looking forward to multiple
file uploads) on the `fileSuccess` event.