Include checksum when completing upload.
A checksum (an SHA256 digest) is sent on completion of a multi-part
upload.  Since Resumable doesn't make the raw file data available in a
`ResumableChunk` we have to resort to reading the file independently
before the chunks are transmitted.

Having to read the entire file is an achitectural oddity; ideally we'd
want to update the accumulator in the digester as each chunk becomes
ready for transmission.  Unfortunately because of bug in the released
version of Resumable we can't use the `preprocess` hook; see
https://github.com/csm-adapt/sagittariidae-fe/issues/4 for details.

;; closes #1