Trivial runtime environment configuration.
This introduces a very lightweight mechanism for managing configuration
for the runtime environment, i.e. 'development' or 'production'.

Since the FE code executes in an environment that we don't
control, viz. the client browser, we can't infer the deployment context
from that environment.  Instead then, we build the source as appropriate
for the desired deployment context.