Add non-JS resources from CLJSJS JARs.
Use the `sift` task's `add-jar` option to add non-JS resources into the
fileset that is produced by the `build` task.  The need for this came as
a surprise to me because it wasn't necessary for the `dev` task.  I
think this is because the `serve` task (from pandeiro/boot-http)
automatically includes resources on the classpath when serving files.

There is still an intermittent error with `react-bootstrap.min.js.map`
that I haven't been able to resolve, but which I think is related.

cf. https://github.com/cljsjs/packages/wiki/Non-JS-Assets